requirements.txt
PyNaCl
websockets
gmpy2